Generates useful metadata
Prior to this commit, the generated metadata were mostly empty and did
not really validate what was instructed in the guide.

One reason was that `ClientResources` was treating the underlying types
as mono value object so no meta-data were generated on them. The other
reason was that the concrete type wasn't returned and the interface has
no setter so no properties were found either.

This commit fixes the infrastructure so that the generated metadata is
valid.